requirements.txt
fastapi
uvicorn
dimod
dwave-samplers
trimesh
numpy
scipy
pydantic
python-multipart
pennylane
pennylane-lightning
amazon-braket-pennylane-plugin
